Add chat handler to bot playground API and invalidate history query on successful form submission
