+++ app/ui/src/components/Bot/Playground/Message.tsx
@@ -64,15 +64,11 @@ export const PlaygroundMessage = (props: Props) => {
           </div>
           <div className="relative flex w-[calc(100%-50px)] flex-col gap-1 md:gap-3 lg:w-[calc(100%-115px)]">
             <div className="flex flex-grow flex-col gap-3">
-              <div className="min-h-20 flex flex-col items-start gap-4 whitespace-pre-wrap break-words">
-                <div className="w-full break-words dark:prose-invert dark">
-                  {props.isBot ? (
-                    <Markdown message={props.message} />
-                  ) : (
-                    <p className="text-sm">{props.message}</p>
-                  )}
-                </div>
-              </div>
+              {props.isBot ? (
+                <Markdown message={props.message} />
+              ) : (
+                <p className="text-sm">{props.message}</p>
+              )}
             </div>
 
             {props.isBot && (
