+++ app/ui/src/components/Common/Markdown.tsx
@@ -27,7 +27,7 @@ export default function Markdown({ message }: { message: string }) {
 
   return (
     <ReactMarkdown
-      className={`prose-invert flex-1 `}
+      className="prose break-words dark:prose-invert text-sm prose-p:leading-relaxed prose-pre:p-0"
       remarkPlugins={[remarkGfm, remarkMath]}
       rehypePlugins={[rehypeMathjax]}
       components={{
@@ -96,133 +96,8 @@ export default function Markdown({ message }: { message: string }) {
             </a>
           );
         },
-        li({ node, ...props }) {
-          return (
-            <li
-              className="list-decimal text-sm marker:text-gray-500"
-              {...props}
-            >
-              {props.children}
-            </li>
-          );
-        },
-        ul({ node, ...props }) {
-          return (
-            <ul className="mx-2 text-sm" {...props}>
-              {props.children}
-            </ul>
-          );
-        },
-        ol({ node, ...props }) {
-          return (
-            <ul className="mx-2 text-sm" {...props}>
-              {props.children}
-            </ul>
-          );
-        },
-        h1({ node, ...props }) {
-          return (
-            <h1 className="text-2xl font-bold" {...props}>
-              {props.children}
-            </h1>
-          );
-        },
-        h2({ node, ...props }) {
-          return (
-            <h2 className="text-xl font-bold" {...props}>
-              {props.children}
-            </h2>
-          );
-        },
-        h3({ node, ...props }) {
-          return (
-            <h3 className="text-lg font-bold" {...props}>
-              {props.children}
-            </h3>
-          );
-        },
-        h4({ node, ...props }) {
-          return (
-            <h4 className="text-base font-bold" {...props}>
-              {props.children}
-            </h4>
-          );
-        },
-        h5({ node, ...props }) {
-          return (
-            <h5 className="text-sm font-bold" {...props}>
-              {props.children}
-            </h5>
-          );
-        },
-        h6({ node, ...props }) {
-          return (
-            <h6 className="text-xs font-bold" {...props}>
-              {props.children}
-            </h6>
-          );
-        },
-        p({ node, ...props }) {
-          return (
-            <p className="text-sm" {...props}>
-              {props.children}
-            </p>
-          );
-        },
-        table({ node, ...props }) {
-          return (
-            <table
-              className="border-collapse border text-sm border-gray-300"
-              {...props}
-            >
-              {props.children}
-            </table>
-          );
-        },
-        thead({ node, ...props }) {
-          return (
-            <thead className="bg-gray-100" {...props}>
-              {props.children}
-            </thead>
-          );
-        },
-        tbody({ node, ...props }) {
-          return (
-            <tbody className="divide-y divide-gray-300" {...props}>
-              {props.children}
-            </tbody>
-          );
-        },
-        tr({ node, ...props }) {
-          return (
-            <tr className="divide-x divide-gray-300" {...props}>
-              {props.children}
-            </tr>
-          );
-        },
-
-        th({ node, ...props }) {
-          return (
-            <th className="p-2" {...props}>
-              {props.children}
-            </th>
-          );
-        },
-
-        td({ node, ...props }) {
-          return (
-            <td className="p-2" {...props}>
-              {props.children}
-            </td>
-          );
-        },
-
-        blockquote({ node, ...props }) {
-          return (
-            <blockquote className="border-l-4 border-gray-300 pl-2" {...props}>
-              {props.children}
-            </blockquote>
-          );
+        p({ children }) {
+          return <p className="mb-2 last:mb-0">{children}</p>;
         },
       }}
     >
