+++ server/src/routes/api/v1/bot/playground/handlers/index.ts
@@ -1,2 +1,3 @@
 export * from "./post.handler"
 export * from "./get.handler"
+export * from "./chat.handler"
